move use-design under `hooks` from `query-hooks`
