+++ editor/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from "./use-design";
 export * from "./use-async-effect";
