+++ editor/query-hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./use-design";
+export const dummy = "wip";
